fixing myio_comm_line_get_double

+++ src/myio.c
@@ -106,7 +106,7 @@ int myio_comm_line_get_double(command_line_t *command_line, const char *option_n
   int i = 0; SEARCH_ARGV_INDEX(i, option_name)
   if(i >= command_line->argc - 1) return 1;
 
-  *value = atoi(command_line->argv[i+1]);
+  *value = atof(command_line->argv[i+1]);
   *found = true;
 
   return 0;
